+++ DataAccess/Domain/Transactions/VendorInvoiceTxn/VendorSalesInvoiceEntity.cs
@@ -14,7 +14,7 @@ namespace DataAccess.Domain.Transactions.VendorInvoiceTxn
         public int VendorInvoiceTxnID { get; set; }
         public string type { get; set; }
         public string? invoiceNo { get; set; }
-        public int? amount { get; set; }
+        public decimal? amount { get; set; }
         public string? estimateNo { get; set; }
         public string? remarks { get; set; }
         public string? postedInTally { get; set; }
